fix(web): gate chat behind auth
